Moved Rag namespace to VectorDb
