+++ PencilCase.API/Program.cs
@@ -6,8 +6,8 @@ using PencilCase.API.Handlers;
 using PencilCase.Shared.Data.Database;
 using PencilCase.LLM.Agents.Providers;
 using PencilCase.LLM.Agents.Providers.Gemini;
-using PencilCase.LLM.RAG;
-using PencilCase.LLM.RAG.Providers.Pinecone;
+using PencilCase.LLM.VectorDb;
+using PencilCase.LLM.VectorDb.Providers.Pinecone;
 using PencilCase.Shared.Models.Notebooks;
 using PencilCase.Shared.Models.Telemetry.LLM.Agents;
 using PencilCase.Shared.Models.Telemetry.LLM.RAG;
