+++ PencilCase.Tests.API/Endpoints/LlmApiEndpointsHandlerUnitTests.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Http.HttpResults;
 using Moq;
 using PencilCase.API.Handlers;
 using PencilCase.LLM.Agents.Providers;
-using PencilCase.LLM.RAG;
+using PencilCase.LLM.VectorDb;
 using PencilCase.Shared.Data.Database;
 using PencilCase.Shared.DTOs.Requests.Llm;
 using PencilCase.Shared.DTOs.Requests.Rag;
