+++ PencilCase.Tests.API/Endpoints/LlmApiFunctionalTests.cs
@@ -4,7 +4,7 @@ using Moq;
 using PencilCase.API.Handlers;
 using PencilCase.Tests.API.Helpers;
 using PencilCase.LLM.Agents.Providers;
-using PencilCase.LLM.RAG;
+using PencilCase.LLM.VectorDb;
 using PencilCase.Shared.Data.Database;
 using PencilCase.Shared.DTOs.Requests.Llm;
 using PencilCase.Shared.DTOs.Requests.Rag;
